requirements.txt
fastapi>=0.103.2
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
langchain-anthropic>=0.1.13
langchain-qdrant>=0.1.0
qdrant-client>=1.9.0
fastembed>=0.2.7
redis>=5.0.0
PyMuPDF>=1.23.5
beautifulsoup4>=4.12.2
python-docx>=1.0.1
Pillow>=10.0.1
sentence-transformers>=2.7.0
PyJWT>=2.8.0
slowapi>=0.1.8
ragas>=0.1.7
pandas>=2.1.1
datasets>=2.14.5
open_clip_torch>=2.24.0
torch>=2.0.1
sqlalchemy>=2.0.21
aiohttp>=3.8.5
html2text>=2024.2.26
